integrations/boozelee/Baker-Street-Laboratory-1/framework/mathematical_validation.py
#!/usr/bin/env python3
"""
Baker Street Laboratory - Mathematical Validation Framework
Rigorous mathematical models for validating breakthrough claims

ASSIGNED TO: Qodo Mathematical Analysis Agent
PRIORITY: 1.1 - Statistical Validation Models
TIMELINE: 2-3 days
"""

import numpy as np
import scipy.stats as stats
from typing import Dict, List, Tuple, Optional
from dataclasses import dataclass
import logging
from abc import ABC, abstractmethod

@dataclass
class PerformanceMetrics:
    """Data structure for tracking performance metrics"""
    baseline_performance: float
    collaborative_performance: float
    error_resolution_rate: float
    breakthrough_probability: float
    sample_size: int
    confidence_level: float = 0.95

class CollaborativeImprovementValidator:
    """
    Validate the claimed 13.7% collaborative improvement with statistical rigor
    
    MATHEMATICAL FOUNDATION:
    - Improvement Factor = (Collaborative_Performance - Baseline_Performance) / Baseline_Performance
    - Statistical Significance: Two-sample t-test with Bonferroni correction
    - Confidence Intervals: Bootstrap sampling with 95% confidence
    """
    
    def __init__(self, significance_level: float = 0.05):
        self.significance_level = significance_level
        self.logger = logging.getLogger(__name__)
    
    def calculate_improvement_factor(self, 
                                   baseline_samples: List[float], 
                                   collaborative_samples: List[float]) -> Dict[str, float]:
        """
        Calculate collaborative improvement factor with statistical validation
        
        Args:
            baseline_samples: Performance measurements from isolated agents
            collaborative_samples: Performance measurements from collaborative agents
            
        Returns:
            Dictionary containing improvement factor, p-value, and confidence interval
        """
        # TODO for Qodo Agent: Implement statistical analysis
        # 1. Calculate mean improvement factor
        # 2. Perform two-sample t-test for significance
        # 3. Calculate confidence intervals using bootstrap
        # 4. Validate against claimed 13.7% improvement
        
        baseline_mean = np.mean(baseline_samples)
        collaborative_mean = np.mean(collaborative_samples)
        
        # Placeholder implementation - REPLACE WITH REAL STATISTICS
        improvement_factor = (collaborative_mean - baseline_mean) / baseline_mean
        
        # TODO: Implement proper statistical testing
        t_statistic, p_value = stats.ttest_ind(collaborative_samples, baseline_samples)
        
        return {
            "improvement_factor": improvement_factor,
            "improvement_percentage": improvement_factor * 100,
            "p_value": p_value,
            "is_significant": p_value < self.significance_level,
            "confidence_interval": self._calculate_confidence_interval(
                baseline_samples, collaborative_samples
            ),
            "sample_sizes": {
                "baseline": len(baseline_samples),
                "collaborative": len(collaborative_samples)
            }
        }
    
    def _calculate_confidence_interval(self, 
                                     baseline_samples: List[float], 
                                     collaborative_samples: List[float]) -> Tuple[float, float]:
        """
        Calculate confidence interval for improvement factor using bootstrap
        
        TODO for Qodo Agent: Implement bootstrap confidence interval calculation
        """
        # Placeholder - implement bootstrap sampling
        return (0.10, 0.16)  # Example: 10% to 16% improvement range
    
    def validate_breakthrough_probability(self, 
                                        research_metrics: List[PerformanceMetrics]) -> Dict[str, float]:
        """
        Validate breakthrough probability calculations with Bayesian analysis
        
        MATHEMATICAL FOUNDATION:
        - Prior probability based on historical research success rates
        - Likelihood function based on current system performance
        - Posterior probability using Bayes' theorem
        
        TODO for Qodo Agent: Implement Bayesian probability model
        """
        # Placeholder implementation
        return {
            "prior_probability": 0.15,  # Historical research breakthrough rate
            "likelihood": 0.85,         # System performance likelihood
            "posterior_probability": 0.75,  # Calculated breakthrough probability
            "credible_interval": (0.65, 0.85)  # 95% credible interval
        }

class ErrorResolutionMetrics:
    """
    Validate the claimed 95% automated error resolution rate
    
    MATHEMATICAL FOUNDATION:
    - Resolution Rate = Resolved_Errors / Total_Errors
    - Confidence Interval: Wilson score interval for proportions
    - Trend Analysis: Time series analysis with seasonal decomposition
    """
    
    def __init__(self):
        self.logger = logging.getLogger(__name__)
    
    def track_error_resolution_rate(self, 
                                  error_log: List[Dict], 
                                  resolution_log: List[Dict]) -> Dict[str, float]:
        """
        Calculate actual error resolution rate with statistical validation
        
        Args:
            error_log: List of error events with timestamps and types
            resolution_log: List of resolution events with timestamps and methods
            
        Returns:
            Dictionary containing resolution rate, confidence interval, and trend analysis
            
        TODO for Qodo Agent: Implement comprehensive error analysis
        """
        # Placeholder implementation
        total_errors = len(error_log)
        resolved_errors = len(resolution_log)
        
        if total_errors == 0:
            return {"error": "No errors to analyze"}
        
        resolution_rate = resolved_errors / total_errors
        
        # TODO: Implement Wilson score interval for confidence
        confidence_interval = self._wilson_score_interval(resolved_errors, total_errors)
        
        return {
            "resolution_rate": resolution_rate,
            "resolution_percentage": resolution_rate * 100,
            "total_errors": total_errors,
            "resolved_errors": resolved_errors,
            "confidence_interval": confidence_interval,
            "meets_target": resolution_rate >= 0.95,
            "trend_analysis": self._analyze_resolution_trends(error_log, resolution_log)
        }
    
    def _wilson_score_interval(self, successes: int, trials: int, 
                              confidence: float = 0.95) -> Tuple[float, float]:
        """
        Calculate Wilson score interval for proportion confidence interval
        
        TODO for Qodo Agent: Implement Wilson score interval calculation
        """
        # Placeholder - implement proper Wilson score interval
        return (0.92, 0.98)  # Example confidence interval
    
    def _analyze_resolution_trends(self, 
                                 error_log: List[Dict], 
                                 resolution_log: List[Dict]) -> Dict[str, float]:
        """
        Analyze trends in error resolution over time
        
        TODO for Qodo Agent: Implement time series trend analysis
        """
        # Placeholder implementation
        return {
            "trend_slope": 0.02,  # Improving by 2% per month
            "seasonal_component": 0.05,  # 5% seasonal variation
            "forecast_next_month": 0.96  # Predicted resolution rate
        }

class PerformanceBenchmarkValidator:
    """
    Comprehensive performance validation for all breakthrough claims
    
    TODO for Qodo Agent: Implement comprehensive benchmarking framework
    """
    
    def __init__(self):
        self.logger = logging.getLogger(__name__)
        self.collaborative_validator = CollaborativeImprovementValidator()
        self.error_metrics = ErrorResolutionMetrics()
    
    def validate_all_claims(self, 
                          performance_data: Dict[str, List[float]]) -> Dict[str, Dict]:
        """
        Validate all breakthrough claims with statistical rigor
        
        TODO for Qodo Agent: Implement comprehensive validation suite
        """
        results = {}
        
        # Validate collaborative improvement
        if "baseline" in performance_data and "collaborative" in performance_data:
            results["collaborative_improvement"] = self.collaborative_validator.calculate_improvement_factor(
                performance_data["baseline"],
                performance_data["collaborative"]
            )
        
        # TODO: Add validation for other claims
        # - Speed improvements beyond 10x
        # - Efficiency improvements beyond 20%
        # - Breakthrough probability accuracy
        # - System scalability metrics
        
        return results
    
    def generate_validation_report(self, validation_results: Dict) -> str:
        """
        Generate comprehensive validation report
        
        TODO for Qodo Agent: Implement detailed reporting system
        """
        # Placeholder implementation
        return "Comprehensive validation report with statistical analysis"

# Example usage and testing framework
def main():
    """
    Example usage of mathematical validation framework
    
    TODO for Qodo Agent: Implement comprehensive testing examples
    """
    # Example data for testing
    baseline_performance = [0.75, 0.78, 0.72, 0.76, 0.74]  # Isolated agent performance
    collaborative_performance = [0.85, 0.88, 0.82, 0.86, 0.84]  # Collaborative performance
    
    validator = CollaborativeImprovementValidator()
    results = validator.calculate_improvement_factor(baseline_performance, collaborative_performance)
    
    print("🔬 Baker Street Laboratory - Mathematical Validation Results:")
    print(f"Improvement Factor: {results['improvement_percentage']:.1f}%")
    print(f"Statistical Significance: {results['is_significant']}")
    print(f"P-value: {results['p_value']:.4f}")

if __name__ == "__main__":
    logging.basicConfig(level=logging.INFO)
    main()
