integrations/boozelee/Baker-Street-Laboratory-1/requirements.txt
# Baker Street Laboratory - Python Dependencies (Python 3.13 Compatible)
# Core AI and ML Libraries
openai>=1.0.0
python-dotenv>=1.0.0
pyyaml>=6.0
pydantic>=2.0.0

# Data Processing and Analysis
pandas>=2.0.0
numpy>=1.24.0

# Web Scraping and API Integration
requests>=2.31.0
beautifulsoup4>=4.12.0
aiohttp>=3.8.0

# Configuration and Environment
click>=8.1.0
tqdm>=4.65.0

# Testing and Quality
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Logging and Monitoring
rich>=13.0.0

# Database
sqlalchemy>=2.0.0

# Async and Concurrency
aiofiles>=23.0.0

# Utilities
markdown>=3.4.0

# API Framework
flask>=3.0.0
flask-cors>=4.0.0
flask-restx>=1.3.0
fastapi>=0.104.0
uvicorn>=0.24.0
gunicorn>=21.2.0

# Production API Management
flask-limiter>=3.5.0
flask-caching>=2.1.0
redis>=5.0.0
celery>=5.3.0

# Cloud & Monitoring
prometheus-flask-exporter>=0.23.0
sentry-sdk[flask]>=1.38.0
python-dotenv>=1.0.0

# Additional APIs Integration
wikipedia>=1.4.0
arxiv>=2.1.0
newsapi-python>=0.2.7
tweepy>=4.14.0
google-search-results>=2.4.2

# Local AI Integration
ollama>=0.1.7
httpx>=0.25.0
